Partition.InitializeAsync exceptions improvements.
If a partition initializes with an exception we now fail all the tests that would have run fast.
These tests will all fail with the exception of IPartionLifetime.InitializeAsync.

Prior to this change these tests would not run at all and the exception would be printed to console.

Okay for CI purposes but in IDE's you want to see the tests fail properly too.

IPartition can now also provide failure text to failed tests so debugging becomes easier in IDE's too.

Lastly the filters now mark tests as skipped rather than invalid. This ensures the tests will always be displayed
and reported.